esdm_es: extract shared timestamp handling logic code
